requirements.txt
flask
flask_cors
fastai
torch
base64
pillow